refactor: Default rich links to fix storybook arrors
